requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-crispy-forms==1.14.0
django-rosetta==0.9.8
django-tinymce==3.4.0
idna==3.3
Pillow==9.2.0
polib==1.1.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
